Bug #327164 Coverity85883
[bug number  ] 327164
[root cause  ] code defect
[changes     ] fix
[side effects] n/a
[self test   ] ok
[whether AOB ]
[reviewers   ] yanbin.fang,mengya.liu

+++ tools/sprd_res_monitor/oprofile/oprofiledaemon.h
@@ -9,7 +9,7 @@ typedef enum
 typedef struct profile_info
 {
     profilecmd cmd;
-    int profiletime;
+    unsigned long profiletime;
 }profileinfo;
 
 #ifdef __cplusplus
